# docs/swagger.yaml
openapi: 3.0.3
info:
  title: CrowdFlow Event Ticketing API
  description: |
    API documentation for the CrowdFlow Venue and Event Management platform.

    ### Security Architecture:
    - **Session Authentication**: A short-lived JWT access token is set as a secure `HttpOnly` cookie named `access_token` (Path `/`). It is paired with a long-lived, rotating, server-revocable `refresh_token` cookie (scoped to Path `/api/v1/auth`) which `POST /api/v1/auth/refresh` exchanges for a fresh access token once the access token expires.
    - **CSRF Protection**: State-changing endpoints require Double-Submit Cookie verification. The client extracts the value of the `csrf_token` cookie and submits it in the custom header `X-CSRF-Token`. Only `/api/v1/auth/login` and `/api/v1/auth/register` are exempt.
    - **Admin Console**: Every `/api/v1/admin/*` route requires the `Super Admin` platform role.

    All responses (success and error) are wrapped in a standard JSON response envelope.
  version: 1.0.0
servers:
  - url: http://localhost:8080
    description: Development Server
  - url: /
    description: Proxy Gateway (Nginx)
paths:
...
  /api/v1/events:
    get:
      summary: List Events (Public Discovery)
      description: |
        Retrieves a paginated list of events. Authentication is optional and does not change the result set.
      parameters:
        - name: limit
          in: query
          required: false
          description: Page size. Rounded up to the nearest multiple of 10 and capped at 100 (e.g. 7â†’10, 101â†’100). Defaults to 20 when absent or non-integer.
          schema:
            type: integer
            default: 20
            maximum: 100
        - name: offset
          in: query
          required: false
          description: Row offset. Floored to the nearest multiple of 10 (e.g. 15â†’10); negative values clamp to 0. Defaults to 0.
          schema:
            type: integer
            default: 0
      responses:
        '200':
          description: List retrieved successfully.
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/StandardResponse'
                  - type: object
                    properties:
                      data:
                        type: array
                        items:
                          $ref: '#/components/schemas/EventListResponse'
        '500':
          description: Database/Internal error (`INTERNAL_ERROR`).
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    post:
      summary: Create Event
      description: |
        Creates a new event from a multipart form. Requires the Event Organizer platform role
        (Super Admin bypasses the check and may additionally set `organizer_id` in `event_data`
        to create on behalf of another organizer; for non-admins `organizer_id` is always forced
        to the caller's own ID).

        Total request body is limited to 10MB.
      security:
        - CookieAuth: []
        - CSRFToken: []
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required:
                - event_data
              properties:
                event_data:
                  type: string
                  description: JSON-encoded string matching the CreateEventRequest schema.
                  example: '{"venue_id":1,"title":"Symphony Orchestra Concert","description":"An evening of classical symphonies.","starts_at":"2026-10-15T19:30:00Z","ends_at":"2026-10-15T22:00:00Z","entertainment_tax_rate":0.1,"entertainment_tax_passed_to_buyer":true,"event_type_id":1}'
                cover_image:
                  type: string
                  format: binary
                  description: Optional cover image. Validated by MIME signature sniffing; max 10MB.
      responses:
        '201':
          description: Event created successfully (status starts as `draft`).
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/StandardResponse'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/EventDetailResponse'
        '400':
          description: Malformed multipart form, missing/invalid `event_data`, or cover image failed validation (`BAD_REQUEST`, `INVALID_FILE`).
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Session token is missing or expired (`UNAUTHORIZED`).
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Platform role check failed (Event Organizer required) or CSRF mismatch (`FORBIDDEN`).
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '422':
          description: Validation of event properties failed (`VALIDATION_FAILED`).
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Cover image upload to object storage failed (`UPLOAD_FAILED`).
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/v1/events/{id}:
    get:
      summary: Get Event Details
      description: |
        Retrieves detailed parameters of a specific event. Authentication is optional; events that are
        not yet `approved` are only visible to their organizer or a Super Admin - all other callers
        receive a 404 (not a 403, to prevent draft-event ID enumeration).
      parameters:
        - name: id
          in: path
          required: true
          description: The numeric database ID of the event.
          schema:
            type: integer
      responses:
        '200':
          description: Event details retrieved.
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/StandardResponse'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/EventDetailResponse'
        '400':
          description: Event ID missing or not an integer (`INVALID_ID`).
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Event not found, or not approved and caller is not its organizer/Super Admin (`NOT_FOUND`).
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/v1/events/{id}/publish:
    patch:
      summary: Publish Event
      description: Publishes an event. Requires an event-scoped Event Organizer role for this specific event (Super Admin bypasses the check).
      security:
        - CookieAuth: []
        - CSRFToken: []
      parameters:
        - name: id
          in: path
          required: true
          description: The numeric database ID of the event.
          schema:
            type: integer
      responses:
        '200':
          description: Event published successfully.
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/StandardResponse'
                  - type: object
                    properties:
                      data:
                        type: object
                        properties:
                          message:
                            type: string
                            example: Event published successfully
                          event_id:
                            type: integer
                            example: 5
        '400':
          description: Event ID path parameter missing or malformed (`INVALID_ID`, `BAD_REQUEST`).
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Authentication missing (`UNAUTHORIZED`).
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Event role authorization failed (user is not an Event Organizer of this event) or CSRF mismatch (`FORBIDDEN`).
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Publication execution failed (`INTERNAL_ERROR`).
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
  /api/v1/venues:
    get:
      summary: List Venues
      description: Lists all venues. Requires the Event Organizer platform role (Super Admin bypasses the check).
      security:
        - CookieAuth: []
      responses:
        '200':
          description: Venues retrieved.
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/StandardResponse'
                  - type: object
                    properties:
                      data:
                        type: array
                        items:
                          $ref: '#/components/schemas/VenueResponse'
        '401':
          description: Session token missing, invalid, or expired (`UNAUTHORIZED`).
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Platform role check failed - Event Organizer required (`FORBIDDEN`).
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Database/Internal error (`INTERNAL_ERROR`).
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/v1/event-types:
    get:
      summary: List Event Types
      description: Lists all event type categories. Requires the Event Organizer platform role (Super Admin bypasses the check).
      security:
        - CookieAuth: []
      responses:
        '200':
          description: Event types retrieved.
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/StandardResponse'
                  - type: object
                    properties:
                      data:
                        type: array
                        items:
                          $ref: '#/components/schemas/EventTypeResponse'
        '401':
          description: Session token missing, invalid, or expired (`UNAUTHORIZED`).
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Platform role check failed - Event Organizer required (`FORBIDDEN`).
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Database/Internal error (`INTERNAL_ERROR`).
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
components:
  securitySchemes:
    CookieAuth:
      type: apiKey
      in: cookie
      name: access_token
      description: Short-lived JWT access token. Sent on every request (Path `/`).
    RefreshCookie:
      type: apiKey
      in: cookie
      name: refresh_token
      description: |
        Long-lived, rotating, server-revocable refresh token. Scoped to Path
        `/api/v1/auth`, so it is only sent to `/api/v1/auth/refresh` (to rotate)
        and `/api/v1/auth/logout` (to revoke). Used by `POST /api/v1/auth/refresh`
        to mint a new access token when the access token has expired.
    StateCookie:
      type: apiKey
      in: cookie
      name: oauth_state
      description: Google oauth flow verification token.
    CSRFToken:
      type: apiKey
      in: header
      name: X-CSRF-Token
      description: Double-Submit CSRF Token header. Must match values in the `csrf_token` cookie.
  schemas:
    StandardResponse:
      type: object
      required:
        - success
      properties:
        success:
          type: boolean
          description: Set to true if request completed successfully.
          example: true
    APIError:
      type: object
      required:
        - code
        - message
      properties:
        code:
          type: string
          description: Machine-readable short error classification code.
          example: FORBIDDEN
        message:
          type: string
          description: Human-readable error description context.
          example: You do not have permission to access this resource
        details:
          type: object
          description: Optional key-value structured details about validation failures.
    ErrorResponse:
      type: object
      required:
        - success
        - error
      properties:
        success:
          type: boolean
          example: false
        error:
          $ref: '#/components/schemas/APIError'
...
    VenueResponse:
      type: object
      properties:
        venue_id:
          type: integer
          example: 1
        name:
          type: string
          example: Jakarta Concert Hall
        address:
          type: string
          example: Jl. Sudirman No. 1
        city:
          type: string
          example: Jakarta
        province:
          type: string
          example: DKI Jakarta
        total_capacity:
          type: integer
          example: 500
    EventTypeResponse:
      type: object
      properties:
        event_type_id:
          type: integer
          example: 1
        event_type:
          type: string
          example: Concert
    OrganizerResponse:
      type: object
      properties:
        organizer_id:
          type: integer
          example: 12
        name:
          type: string
          example: Richie Obhasa
        avatar_url:
          type: string
          example: https://cdn.crowdflow.example/avatars/12.jpg
        profile_url:
          type: string
          example: /organizers/12
    EventListResponse:
      type: object
      properties:
        event_id:
          type: integer
          example: 5
        title:
          type: string
          example: Symphony Orchestra Concert
        starts_at:
          type: string
          format: date-time
          example: "2026-10-15T19:30:00Z"
        ends_at:
          type: string
          format: date-time
          example: "2026-10-15T22:00:00Z"
        category:
          type: string
          enum: [concert, festival, sport, conference, exhibition, community, workshop_seminar, other]
          example: concert
        cover_image_url:
          type: string
          example: https://cdn.crowdflow.example/events/covers/1720000000000000000.jpg
        venue:
          $ref: '#/components/schemas/VenueResponse'
        organizer:
          $ref: '#/components/schemas/OrganizerResponse'
    EventDetailResponse:
      type: object
      properties:
        event_id:
          type: integer
          example: 5
        title:
          type: string
          example: Symphony Orchestra Concert
        description:
          type: string
          example: An evening featuring classical symphonies from Mozart and Beethoven.
        starts_at:
          type: string
          format: date-time
          example: "2026-10-15T19:30:00Z"
        ends_at:
          type: string
          format: date-time
          example: "2026-10-15T22:00:00Z"
        category:
          type: string
          enum: [concert, festival, sport, conference, exhibition, community, workshop_seminar, other]
          example: concert
        entertainment_tax_rate:
          type: number
          format: double
          example: 0.1
        entertainment_tax_passed_to_buyer:
          type: boolean
          example: true
        cover_image_url:
          type: string
          example: https://cdn.crowdflow.example/events/covers/1720000000000000000.jpg
        venue:
          $ref: '#/components/schemas/VenueResponse'
        organizer:
          $ref: '#/components/schemas/OrganizerResponse'
...